backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
python-jose[cryptography]
passlib[argon2]
argon2-cffi
python-multipart
psutil
Pillow
prometheus-client
slowapi
redis
celery
structlog

pytest
pytest-asyncio
httpx
